test(text_just): add four test cases for text_justification #10
Added four test cases for text justification that increases the coverage
of the function to 100%. The only branch that was not covered before was
the branch of the valueError which s now in the test cases. Also added test
cases for edge-cases, sepcifically when a word is exactly as long as max_width
and also when two words are as long as max_width. The function acts as
expected without bugs.