refactor(three_sum): refactor three_sum to decrease cyc complexity #13 #15
refactored the function three_sum in the array module in order to
decrease the cyclomatic complexity of the functin and also to make
the code shorter and more easily understood. I also simplified the
function and made use of the python set property that there can be
no duplicates in a set, this significantly shortened the code.